temp: add debug data in interest job
I would run '$ bats -t bats/credit-facility-sim-time.bats' and then
observe the different tables in postgres to wait for job execution,
especially job execution state data for interest job.

So far, I can observe that the ledger incurrence eventually gets
persisted but the accrual event is never persisted, and nothing in the
code signals that there was bad or incomplete entity state, or that
the entity persistence operation failed.