chore(reporting): configure incremental updates for large append only sources
Even without real production data, many of the source tables creaed by
`tap-postgres` are already hundreds of thousands of rows long.
This might not be an issue once incremental loading is configured,
but in the meantime this is a simple fix that will prevent running
dbt from taking too long in the future.