feat(reporting): enable live views for holistics
This PR configures the dbt dataset as an authorized dataset
to our raw dataset.  This means that, though the holistics
service account can't query the raw dataset, any views we create
in the dbt dataset which point to the raw dataset can be
queried by holistics.  This means we can get live prices
and other nice things.