chore: In RBAC hierarchy call 'parent' the more powerful role
In casbin model's group definitions, the first subject is the one that
is also a member of the second one. For example 'g, admin, user' means
that role admin also has role user and thus admin is the more powerful
role. In parent/child naming scheme, it is more natural to call the
more powerful role (admin) 'parent' and the less powerful (user)
'child'.

This change brings alignment with this idea. Consequently, the order
of roles in `add_role_hierarchy` method conforms to the order in
casbin model.