chore(reporting): only extract and load required tables from pg
Currently we extract and load every table in the public schema in pg.
That's 121 tables of which we only use ~40 in the transformation pipeline.
This commit specifies only the used tables for import:

	yq '.sources | .[] | .tables | .[] | .name' meltano/transform/models/sources.yml |
	sed 's/_view$//g ; s/^/        - /g ; s/public_/public-/g'

We save some unecessary work when running the elt pipeline but we'll have to
add tables here explicitly if we want to add them to the set used in the dbt
transformations.