chore(reporting): run full ETL on on-demand report generation (#2988)
* fix(reporting): install dbt dependencies before running

The missing dependency (which is only needed for running the tests) was
causing all dbt runs to fail leading to the missing reports.

* chore(reporting): run full ETL on on-demand report generation

We're moving to dagster so this is a temporary fix but
ensuring that reports generated through the ui have the latest
data will make it much easier to review them.